Resume plato's resnet and vgg
